Add GitHub Action workflow to create release post rockstor-core release
We currently manually create a new release of rockstor-jslibs every time
a release in rockstor-core is made; as required by our rockstor-rpmbuild
process.

This commit creates a GitHub Action workflow to automatically do that;
this workflow is meant to be triggered by a parent workflow present in
rockstor-core.

This commit also includes a .gitattributes file to exclude git and
GitHub config files from the tarball of assets automatically created
upon release creation.